shoe-store/requirements.txt
black==23.3.0
coverage==7.2.5
dj-database-url==1.3.0
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-allauth==0.54.0
django-storages==1.13.2
dj-rest-auth==3.0.0
factory-boy==3.2.1
google-cloud-storage==2.8.0
gunicorn==20.1.0
isort==5.12.0
Pillow==9.5.0
pre-commit==3.3.2
PyYAML==6.0
python-dotenv==1.0.0
ruff==0.0.267
uritemplate==4.1.1
